Version 1.6 with new MCSplus

+++ src/main/java/org/openscience/smsd/algorithm/mcgregor/QueryProcessor.java
@@ -239,11 +239,11 @@ public class QueryProcessor {
         int corresponding_atom = 0;
         for (int a = 0; a < mapped_atoms_size; a++) {
             if ((molecule == 1)
-                    && (mapped_atoms.get(a * 2 + 0).intValue() == atom_from_other_molecule)) {
+                    && (mapped_atoms.get(a * 2 + 0) == atom_from_other_molecule)) {
                 corresponding_atom = mapped_atoms.get(a * 2 + 1);
             }
             if ((molecule == 2)
-                    && (mapped_atoms.get(a * 2 + 1).intValue() == atom_from_other_molecule)) {
+                    && (mapped_atoms.get(a * 2 + 1) == atom_from_other_molecule)) {
                 corresponding_atom = mapped_atoms.get(a * 2 + 0);
             }
         }
